fix (almost) scroll down behaviour on last page
the logic of the previous commit wasn't correct, trying too hard
to not display a black border after the last page, it effectively
prevented one from actually scrolling down to the very end of the
document depending on page size, screen size and zoom.
it was however effective in preventing from being able to endlessly
scroll down and having the last page re-rendered over and over.

the new logic gets it almost right - if the screen height is smaller
than the last 2 pages combined, then no black border is displayed
anymore after the last page and scrolling is stopped precisely on
the border. if otoh the screen size is bigger than the last 2 pages
combined, the border is still partially visible.
however, sdlbook was never designed to show more than 2 pages so
other odd behaviour occurs as well, but it's no big deal, nobody
reads ebooks with a magnifying glass.

the goto is used in case the last page was already visible, but not
set as the current page yet. without it, a page down press would
then get into the black area again.