fix glibc crash in isdigit() using a custom impl  (#8)
GLIBC's ctype functions crash when called with any value other
than 0-255 or EOF[0].
the SDL keysym could contain values outside of that range.
it's not sufficient to simply truncate the value, as it could lead
to values meant for another purpose to be detected as digit.

commit message written by commiter.

[0]: https://drewdevault.com/2020/09/25/A-story-of-two-libcs.html